Move goAsync function to ReceiverExtensions file
